Modified API calls to reflect address list being changed.
Based on modifications to index.js, the mods here receive the address list name in a the 'list' parameter and modify the REST API url to reflect the list being changed.
In original code the variable "addressList" was predefined and fixed.  Here is is modified based on the passed parameter by each function call.