Pull user group info from APM session variables and send to ILXServer node.js plugin
The ILX::call between TCL and the node.js plugin is modified to include one additional parameter.  This parameter is the address list name which reflects the user group name.  The group name in pulled from the APM session variable "session.cust.user.group".  Use the APM VPE agents to create and populate this variable.  It can contain a fixed value or value derived from the results of an LDAP/AD query.